registration_utils/requirements.txt
click
pyyaml
requests
dtool_lookup_server
